Fixes compile time warnings. A bit of code cleanup.

+++ src/geometry/ChCTriangleMeshConnected.h
@@ -101,7 +101,7 @@ public:
 			/// This is disconnected - no vertex sharing is used even if it could be..
 	virtual void addTriangle(const ChVector<>& vertex0, const ChVector<>& vertex1, const ChVector<>& vertex2)
 	{
-		int base_v = m_vertices.size();
+		size_t base_v = m_vertices.size();
 		m_vertices.push_back(vertex0);
 		m_vertices.push_back(vertex1);
 		m_vertices.push_back(vertex2);
@@ -110,7 +110,7 @@ public:
 			/// Add a triangle to this triangle mesh, by specifying a ChTriangle 
 	virtual void addTriangle(const ChTriangle& atriangle)
 	{
-		int base_v = m_vertices.size();
+		size_t base_v = m_vertices.size();
 		m_vertices.push_back(atriangle.p1);
 		m_vertices.push_back(atriangle.p2);
 		m_vertices.push_back(atriangle.p3);
